smth

+++ shelter-service/src/main/java/com/bme/shelterservice/repository/ShelterRepo.java
@@ -5,6 +5,6 @@ import com.bme.shelterservice.dto.Shelter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ShelterRepo extends JpaRepository<Shelter, Long> {
-		
+public interface ShelterRepo extends JpaRepository<Shelter, Integer> {
+
 }
